service/bot/requirements.txt
aiogram==3.*
asyncpg
httpx
python-dotenv